debaml P3: static output-format parity — reachability ordering + alias/description + BAML oracle (slice 3) (#586)
Slice 3 of the de-BAML P3 native parser. Makes the native static-schema builder
render ctx.output_format BYTE-EXACT vs BAML v0.223.

- Reachability ordering: the builder emits schemadescriptor.Bundle enum/class
  slices in BAML output-format hoist order (reverse-of-reference DFS from the
  target), not declaration/discovery order. Reuses the dynamic path's proven
  traversal via a new (*schema.Bundle).ReachableOrder() wrapping the existing
  orderByReachability, so the two paths share one BAML-parity source of truth.
- @alias -> Name.Alias and @description -> Description lowered on class fields,
  enum values, and class-level block attributes; every other attribute
  (@assert/@check/@stream.*/@@dynamic/@skip/unknown) declines fail-closed.
- BAML-oracle byte-parity harness: a dedicated parity_baml_src project whose
  functions have a prompt of ONLY {{ ctx.output_format }}; a dedicated
  BAML_REST_USE_DEBAML=false container (in SUBPROCESS mode) renders each via the
  real BAML v0.223 runtime (oracle); the same .baml bytes are parsed natively,
  built+ordered into a descriptor, rendered, and byte-compared. Covers string /
  enum / literal-union map keys (with key-then-value reachability pinned by
  ParityEnumKeyMap); the subprocess worker isolates BAML's generated-Decode panic
  on enum/literal-union map keys so the captured render is still byte-compared.
  The expected function set is asserted independently and the proof is pinned to
  the exact 0.223.0 oracle.

Extracted the slice-2 builder from cmd/introspect (package main, not importable)
into internal/nativeschema so both the introspect pipeline and the harness drive
it; regenerated root embed.go (idempotent). Config extraction + existing goldens
byte-identical.

Lax/decline decisions (also logged on #586): @alias/@description args must be a
single plain string (D5); /// doc comments are detected and DECLINE fail-closed
until doc-comment capture lands — not silently dropped (D6); enum-level
@@description declines (no rendered representation); parity corpus returns
class/class[]/primitive only because BAML's Go generator emits invalid code for a
bare map<...> return (maps/unions/optionals tested as fields).
