+++ bamlutils/bamlparser/bamlparser.go
@@ -376,7 +376,9 @@ func (tb *TypeBlock) Parse(lex *lexer.PeekingLexer) error {
 	default:
 		return participle.NextMatch
 	}
-	tb.Keyword = lex.Next().Value
+	kwTok := lex.Next()
+	tb.Keyword = kwTok.Value
+	tb.Span = tokSpan(kwTok)
 
 	if t := lex.Peek(); t.Type == identType {
 		tb.Name = lex.Next().Value
