+++ integration/testdata/parity_baml_src/generators.baml
@@ -1,0 +1,8 @@
+// Generator configuration for the parity fixtures.
+// Note: This gets removed during the build process - only the version is used for detection.
+generator target {
+    output_type "go"
+    output_dir "../"
+    version "0.214.0"
+    default_client_mode async
+}
